requirements.txt
langchain
langgraph
langchain-community
langchain-openai
qdrant-client
pypdf
python-dotenv